chore: split utils, hooks folder and add data folder
